Move modal fragments where they belong
